// src/app/interfaces/player-interface.ts
export interface Player {
    id?:        number;
    firstName?:  string;
    lastName?:   string;
    image?:      string;
    attack?:     number;
    defense?:    number;
    skills?:     number;
    idAuthor?:   number;
    idPosition?: number;
}

export interface BodyPosition {
    id:          number;
    description: string;
}